Includes cobra and flags for the configuration
