+++ go.mod
@@ -4,5 +4,7 @@ go 1.13
 
 require (
 	github.com/rjeczalik/notify v0.9.2
+	github.com/spf13/cobra v0.0.5
+	golang.org/x/sys v0.0.0-20191010194322-b09406accb47 // indirect
 	gopkg.in/yaml.v2 v2.2.4
 )
